requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
idna==3.3
joblib==1.1.0
nltk==3.7
pymediawiki==0.7.1
pywikibot==6.6.5
ratelimit==2.2.1
regex==2022.1.18
requests==2.27.1
soupsieve==2.3.1
tqdm==4.62.3
urllib3==1.26.8
wcwidth==0.2.5
Wikipedia-API==0.5.4
wikitextparser==0.48.0
